Diffusion and model updates

+++ eir/setup/input_setup_modules/setup_array.py
@@ -4,7 +4,6 @@ from typing import Literal, Optional
 
 import numpy as np
 import torch
-from aislib.misc_utils import get_logger
 
 from eir.data_load.data_source_modules.deeplake_ops import (
     get_deeplake_input_source_iterable,
@@ -23,6 +22,7 @@ from eir.setup.setup_utils import (
     al_collector_classes,
     collect_stats,
 )
+from eir.utils.logging import get_logger
 
 logger = get_logger(name=__name__)
 
@@ -95,7 +95,7 @@ def get_array_normalization_values(
         ds_iter = get_deeplake_input_source_iterable(
             deeplake_dataset=deeplake_ds, inner_key=deeplake_inner_key
         )
-        tensor_iterator = (torch.from_numpy(i.numpy()) for i in ds_iter)
+        tensor_iterator = (torch.from_numpy(i.numpy()).float() for i in ds_iter)
     else:
         file_iterator = Path(input_source).rglob("*")
         np_iterator = (np.load(str(i)) for i in file_iterator)
